remove old Lens/Namespace implementations
