+++ code/ext/ext.lib.automerge/src/Doc.Lens.NEXT/Lens.Events.ts
@@ -11,7 +11,7 @@ export function eventsFactory<L extends {}>(
   const { dispose, dispose$ } = life;
   const $ = source$.pipe(rx.takeUntil(dispose$));
 
-  const api: t.LensEvents2<L> = {
+  const api: t.LensEvents<L> = {
     $,
     changed$: rx.payload<t.LensChangedEvent<L>>($, 'crdt:lens:changed'),
     deleted$: rx.payload<t.LensDeletedEvent<L>>($, 'crdt:lens:deleted'),
