--- code/ext/ext.lib.automerge/src/Doc.Lens/Lens.ts
@@ -1,10 +1,0 @@
-import { init } from './Lens.impl';
-import { Registry } from './Lens.Registry';
-
-/**
- * Lens for operating on a sub-tree within a CRDT.
- */
-export const Lens = {
-  init,
-  Registry,
-} as const;
