--- code/ext/ext.lib.automerge/src/Doc.Lens/common.ts
@@ -1,1 +1,0 @@
-export * from '../common';
