--- code/ext/ext.lib.automerge/src/Doc.Lens/index.ts
@@ -1,1 +1,0 @@
-export * from './Lens';
