--- code/ext/ext.lib.automerge/src/Doc.Namespace/common.ts
@@ -1,2 +1,0 @@
-export * from '../common';
-export { init as Lens } from '../Doc.Lens/Lens.impl';
