--- code/ext/ext.lib.automerge/src/Doc.Namespace/index.ts
@@ -1,1 +1,0 @@
-export * from './Namespace';
