+++ code/ext/ext.lib.automerge/src/Doc/Doc.ts
@@ -1,7 +1,5 @@
-import { Lens } from '../Doc.Lens';
-import { Lens as Lens2 } from '../Doc.Lens.NEXT';
-import { Namespace } from '../Doc.Namespace';
-import { Namespace as Namespace2 } from '../Doc.Namespace.NEXT';
+import { Lens } from '../Doc.Lens.NEXT';
+import { Namespace } from '../Doc.Namespace.NEXT';
 import { DocMeta as Meta } from './Doc.Meta';
 import { DocPatch as Patch } from './Doc.Patch';
 import { del } from './Doc.u.delete';
@@ -15,17 +13,11 @@ export const Doc = {
   Meta,
   Data,
   Patch,
+
   Lens,
   Namespace,
-
-  Lens2, // TEMP 🐷
-  Namespace2, // TEMP 🐷
-
   lens: Lens.init,
   namespace: Namespace.init,
-
-  lens2: Lens2.init, // TEMP 🐷
-  namespace2: Namespace2.init, // TEMP 🐷
 
   toObject,
   get,
