+++ code/ext/ext.lib.automerge/src/types.ts
@@ -1,7 +1,5 @@
 export type * from './Doc/t';
-export type * from './Doc.Lens/t';
 export type * from './Doc.Lens.NEXT/t';
-export type * from './Doc.Namespace/t';
 export type * from './Doc.Namespace.NEXT/t';
 
 export type * from './Store.Index/t';
